(related: tdf#133487) sw: fix ordering of virtual SdrObjects for textboxes
Calling XShapes3::sort() on export of the testTdf130314 fails because of
2 consecutive textboxes; the function requires a textbox to immediately
follow its shape in the list (i.e. textbox has OrdNum of shape + 1).

This is because for shapes in header/footer, one virtual SdrVirtObj is
created per page where the header/footer is shown, and the
SwFlyDrawContact::GetOrdNumForNewRef() does not take textbox ordering
into account.

It's not clear if the assumption that the shape's SdrVirtObj is created
before the textbox's always holds, but let's try this for now.

Change-Id: I860896471211bf6c142ab825f298f4d4c0eec148
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112029
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>