do not disable Writer font caching on zoom mismatch
This code disables font caching for some cases, e.g. when exporting
to PDF if the zoom is set to anything else than 100%, or also when
loading a document.

It unfortunately comes from the OOo initial import, so it's
impossible to find out why it should be needed. SwFntObj instances
are created only using SwFntAccess ctor, which checks any SwFntObj
it tries to reuse, so if this some additional factor to consider
when reusing SwFntObj, then it can be checked there. But when I
remove the code, I cannot see any breakage. Tests pass, normal
rendering is fine, document thumbnail is fine, printing is fine,
PDF export is fine, embedding in a spreadsheet is fine.

We had a look at this with Miklos and concluded that it is not
needed (until proven otherwise, in which case it'll be also known
why it would be needed).

Change-Id: I59d375229a265a4353c1da6dd067d892dd4ecfb6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112589
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>