CppunitTest_sw_ooxmlexport: clean up testEffectExtentMargin
The shape had line information, but the intent was to test what happens
when the shape has some non-zero effect extent, line information is not
needed for that.

And do the same for a bunch of other tests: a test document should
ideally have line information only if it's relevant for a test.

Change-Id: I9ea1f38841a30e7e61088d347443e37948b1a9d2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112401
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins