tdf#135774 Char highlight: partial revert tdf#114799
...DOCX import: fix missing numbering highlight
from 7.0 commit 3774fa516953c1e24c6f7805e0b2f88131e1b3a8.

If the numbering had a specified w:highlight in numbering.xml,
then that character background should apply to the bullet point.

However, SetDiffFnt automatically clears the highlight,
the same as it does with mxBackColor (fixed in earlier commit).
So a special step is needed to check if the font has a highlight,
and since that direct formatting has priority, re-apply it
after SetDiffFnt has finished.

This is slightly different from mxBackColor - which is an
optional attribute. Highlight is reset to COL_TRANSPARENT.
If a RES_CHRATR_HIGHLIGHT exists explicitly setting a
COL_TRANSPARENT, then any overrides should have been removed from
pCleanedSet, and so it never needs to SetHighlightColor(AUTO).

Change-Id: Id9d88054254563706b9aee90a1d257b467139334
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112410
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>