devtools: change DocumentModelTreeEntry to do expanding and name
This changes DocumentModelTreeEntry and its subclasses to handle
when to show the expander on the tree view and when not, and to
hold the string of the tree view node, which is passed as a
input parameter. This simplifies the code and makes it easier
to handle. Most of the ideas are from ObjectInspectorNodeInterface
and it subclasses, where this works really good.

Change-Id: I076f80a7c64abd0eb67af7ac0144082611d10076
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112578
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>