Don't use the macOS system German spelling dictionary for Swiss German
The system German dictionary accepts 'ß' which is not used in Swiss
Standard German but completely replaced with 'ss'. We assume that the
LibreOffice German dictionary is installed, too. It has correct
spelling for de_CH.

It would be good if we knew in the macosxspell code whether the
LibreOffice German dictionary installed, and act correspondingly. But
figuring that out seems more complicated than what I am willing to
spend on this for now.

Change-Id: I380d7734cea34a2408d9ed20a8c3d9efe6a8fea2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/115690
Tested-by: Jenkins
Reviewed-by: Tor Lillqvist <tml@collabora.com>