sw reqif-xhtml export: make sure OLE1 preview is WMF-only
reqif-xhtml has previews for embedded objects at 4 levels:

- png preview
- RTF preview
- OLE1 preview
- OLE2 preview

The OLE1 preview has to be a WMF one, and in case our WMF export tries
to inject EMF into it, Word will refuse to open the embedded object. So
add a new flag to allow opting out of the EMF embedding for reqif-xhtml
export purposes.

The other option would be what Word does to just omit the OLE1 preview,
but then this would break reqif-xhtml consumers who take the preview
from the OLE1 data.

Change-Id: Ia8d4626aefb6309743c9f4023f289c9a7b872035
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112648
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins