sw change tracking: antialias deleted comments
Deleted comments get cross out lines with
change tracking, now with (optional) antialiasing
(as in the case of connecting lines of the comments).

Also adjust bottom line ends (which were
removed by the bottom text label of the
comment box, resulting a gap between the
bottom line end and the left/right frame
borders of the comment box).

Change-Id: I73f944ad9b82562d62029617cb029ca396b74ab9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112650
Tested-by: Jenkins
Reviewed-by: László Németh <nemeth@numbertext.org>