VCL force Scheduler instant timeout on task skip
If we skipped the current task processing, we need to immediately
schedule a new timeout, as we know there is a pending task. So set
nMinPeriod to 0, as it contains the period to the next pending
task, and update the system timer.

Fix regression from commit bdbc6ce51eaa5f367a5ebfb596d3d4d9a4320726
("invoke idle priority timers only when actually idle").

Change-Id: Ie3fc176fade595ec7406fec0835835bd86dfbd90
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112762
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>