devtools: don't store locally the object as XIndexAccess
This doesn't correctly read the IndexAccess if we remember it into
a member variable, because we can't rely that the input object is
cast to XIndexAccess, but use a getMainObject method to get a
different object and cast that to XIndexAccess. So the instance
of the object can change (object can be removed or and a new one
added) and we have no way of knowing what the implementation does.

It is possible to cache the XIndexAccess object into a member
variable (if done carefully), but it is prone to the described
problems so let's not remember any intermediate objects for
all DocumentModelTreeEntry subclasses unless we actually have a
performance problem.

Change-Id: I9d9d30d2b43846d473d118d419d407e34d0b1666
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112836
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>