tdf#133933 sw: fix crash on undo of redlined pasted table
This is just a crash fix, there is some deeper problem here around undo:
once redlining is on, overwriting table content during paste does not
delete the fly frames anchored to empty cells (see lcl_CpyBox()), and we
have the same problem on undo as well (see
SwUndoTableCpyTable::UndoImpl()). And node indexes are recorded before
inserting flys, so if they are not deleted, then the indexes don't
match.

Change-Id: I419e96c4d64f70a582358cab3808cea8b0e36649
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112939
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>