tdf#125936 writerfilter: treat escapement in numbering like styles
deferCharacterProperties wasn't occurring in the numbering import
(and so it was affecting the first run of the body text).

But just like character styles, it would be better to just
consider this auto-superscript instead of to defer it
and calculate based on the fontsize - since that really
isn't known until layout time, and so only works with
direct formating.

Change-Id: I9ce5a31c173089603316f4c3389e5f2e5dbe165a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112982
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>