tdf#140336 ms formats export: export NONE background for ParaBackColor
This fixes a regression from LO 4.2 with tdf#88583's
commit 5a3229e9b3f4ead26185433e95c7298a5881a75d.

It isn't only SOLID that needs to be written,
otherwise there is no way to cancel an inherited background.

And only write out a SOLID if one is specified here.
Do not duplicate a style's SOLID fill into every
sub-style and paragraph. (Yikes)

This fixes DOCX (seriously) and DOC (de-duplication),
but not RTF (which doesn't write out transparent brushes).

Surprisingly, I didn't find any examples
where a NONE was specified that matched an inherited NONE.
But it IS important (probably because of a bug) and also smart
to determine and avoid writing NONE if there is no inheritance.
So then, there is no reason not to do the last step of
checking if the current brush matches the inherited brush.

[That also paves the way for a potential follow-up commit
to try deduplicating SOLID fills as well,
although that likely adds grabbag complications into the mix.
And perhaps deduplication isn't desired either,
in case the user actually wanted to duplicate
(if that is even possible under UI conditions).]

Change-Id: I8958a68c779057d9fc5da74eca50fcf5b41d8b52
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/113105
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>