Outline folding: only hide the fold button when needed
Current behavior is to always hide the fold button on mouse leave which
is usually followed by a delayed showing of the same button. This patch
prevents this hide show by only hiding the button when the button mouse
leave results in leaving the area of the frame the button controls.

Change-Id: Iecb544bfeffb32ac4ba7183719a193a96162c7c9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/113243
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>