UI Selection dialog: reword UI description
The description of Groupedbar Compact variant starts with a general
description about both Groupedbar UI variants (the compact variant and
the full variant), then specify what is special about the compact
variant.  Therefore it should starts with "The Groupedbar interface
provides...", just like the description of Groupedbar Full variant in
the previous line.

Change-Id: I98bec4efce996b7f17461a8bb1504e80d70279c5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/113393
Tested-by: Jenkins
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>