android: Don't require that user presses 'Back' twice to exit
Drop the "Press back again to quit" info shown when pressing
"Back" in the file selection dialog in Android Viewer and
just quit the app right away.

It was originally added in

    commit eeba67e116ae2e6162c135b73aecd094084dea28
    Date:   Wed Jun 10 19:04:22 2015 +0200

        tdf#87434: android: system back key to go one level up

        Added an additional check so back has to be pressed twice on the root
        folder to actually leave the application. It's a check seen in many
        other apps.

but I don't really see any need to bother the user
about pressing "Back" again. Nothing is lost at
this stage when quitting the app, and I haven't
seen anything similar in many current apps myself
(but have rather seen some extra confirmation dialogs
disappear from desktop applications over the last
years).

The original request in tdf#87434 to go one level up
in the directory hierarchy is unaffected by this,
though this only applies for the custom widgets
to browse the file system, which will potentially be
dropped in the future anyway, now that support
for the system file picker has been added in

    commit 0bb167970c99d507d015b76d7b25bd06ad07a3c6
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Mar 19 14:29:36 2021 +0100

        tdf#129833 android: Allow opening files using system file picker

Change-Id: Ib324b7f0b82427b04c7708665ff7492a758eec9b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/113413
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>