ScriptForge (scriptforge.py) UI, Document & Calc classes
Changes in the Basic-Python engine:
- arguments of methods may be 2D arrays
- 1st argument of method may be a Basic object
- properties starting with 'X' may contain UNO objects
- GetProperty may contain 1 argument

Calc class is a subclass of the Document class in Python
(in Basic it is simulated)
A CalcReference class is also created to contain
Range and Sheet objects

Many comments have been reviewed for consistency in SF_Calc,
SF_Form and SF_FormControl.xba
Conversion to DataArrays is improved to process input
arrays of arrays. Date conversions are supported.

Change-Id: Id6254d668ec981b00555e7e277c4d31982c00092
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/113568
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>