make clang plugin check code in headers only once if in PCH
When using Clang PCHs, we know for certain that the content of a PCH
will be used once by the PCH's dedicated source file. So it is not
necessary to let clang plugin check locations coming from a PCH
every time, but just once when compiling that dedicated source.

For starmath's parse.cxx this reduces compilation time 0.94s->0.4s
(0.1s when not using plugins at all), for sc's document.cxx it is
5.9s->5.0s (4.0s without plugins). For reference, without PCHs
the numbers are (with/without plugins) 2.1s/1.9s for parse.cxx
and 11.2s/10.3s for document.cxx.

Change-Id: Ie39787e65d7951187941dcff4899d053da63cbdd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/113817
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>