Tip-of-the-day: improve a tip and add link (related to tdf#139533)
This was changed in
    commit 441c2bec1db149a8e72ec228e9e6c8c3e630826a
    tdf#139533 re-word a tip of the day
and "You can" was added to the beginning of all sentences that start
with "Create...".  However this particular one is phrased as a question,
and "You can create...?" is obviously grammatically incorrect.

Fix this by combining the two sentences of the tip into one, and drop
the unnecessary and inaccurate note in parentheses.

Also add a link to the help page about creating master document.

Change-Id: Ib6e92b2a7772378e4b7b67e40bdd2e2791d4f07a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/113392
Tested-by: Jenkins
Reviewed-by: Seth Chaiklin <sdc.blanco@youmail.dk>