tdf#141540 fix docx import of group or line with rotation
... and fix case wrap 'Square' and 'in Line' with them.
Non-uniform scaling of a rotated shape might produce skew. Such had
happened, when setting group or line to the size contained in
GraphicImport. Avoid it.
Writer has special rules for shape position and marging in case of wrap
'Square' and 'in Line', depending on rotation angle. The patch adds
the needed margins.
The patch changes some unit tests where we now get slightly different
values.
The patch fixes the wrong skew in sample document of tdf#73022.

Change-Id: Ic743790c3fc8b8b10a4324d9e0184ad945cdceb6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/114193
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>