tdf#100961 doc/rtf export: save FIXEDFLD as fldlock
This fixes the 5.0 regression by
commit 2596b70cc94e08e7780bb659b995dd883f536074
    fdo#59886 export fixed date and time fields to docx.
which fixed it for DOCX, but broke it (a plain text dump)
for DOC and RTF formats.

We still can't READ fixed fields properly,
(there is no impact/regression on import at all)
but this at least works for MS Word,
and gets us setup for properly importing.

Change-Id: I8835fce06233a52f1433f02879196234f1355280
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/114061
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>