tdf#79591: No need to use getStr here
312bbef37ceb89b11c07b99476dde216faed829d had abused the C-style strings,
and relied on the names containing zero byte to be truncated on the byte.
However, that would only work for names with zero bytes, not with other
control characters. Additionally, that prevented the initial names from
correct round-trip.

This reverts the older fix, and makes sure to handle the attributes with
zeroes correctly (using memcpy instead of strncpy).

It also removes several similar unneeded uses of getStr.

Change-Id: I6c52874d99fe9eb9ccbe0c9a9b57e3b51c45a19f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/114040
Tested-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>