sw define bibliography entry dialog, file picker: handle relative URLs
On one hand, if we have a setting at tools -> options -> load/save ->
general -> Save URLs relative to file system, then let's be consistent
and handle that setting here as well.

On the other hand, this is the default, so this helps the scenario where
a bunch of PDFs are in a directory, a document is next to them, and
copying all of them to a different location doesn't break these links.

Change-Id: Ic6fc552b656a754b29887915ac35ef5c2390f156
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/114049
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins