sw bibliography table, text generator: handle relative URLs
Now that the doc model always contains absolute URLs, it is possible to
decide when it does and when it doesn't make sense to convert to
relative URLs.

The bibliography table does a mixed approach: the document text depends
on configuration (relative by default), but the underlying URL is an
absolute one. This way it's always clickable and the ODT filter can save
it as a relative one (again, depending on configuration).

With this, all fields are meant to be clickable: biblio field vs biblio
table, relative or absolute ones, with or with a (page) fragment, Writer
UI vs PDF export.

Change-Id: Ib6879709314fa127094faf7fa615fe57289b41f1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/114299
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins