NFC: cleanup unnecessary variables, reduce scope.
This function look a LOT different
way back at original import,
and some of the refactoring since then
left things a bit messy.

The impetus for changing this was to see how m_nListLevel
was being used, and having it assigned to a temporary
const variable for no reason was annoying.

Change-Id: Ia25397dbe4bce2fd7bdfbb1b3053bd271c08212a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/114404
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>