Revert "tdf#125949 Allow the slide to continue with ..." (tdf#133447)
These animations were causing problems because of Cairo not handling
well large matrix values. With the real problem fixed this can now
be reverted. I assume the commit claiming to occur with OpenGL was
a mistake, I cannot reproduce any problem, or possibly the problem
went away when the OpenGL VCL backend was removed.

This reverts commit 5aad4e4494822977ce25df1b9ed913a9dba69a84.
This reverts commit 70752cd6abf316c11b44c13fde2d71f716ca0831.

Change-Id: I1c1fff94ff65c7937197a32176a8775ecb2a502e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/114345
Tested-by: Luboš Luňák <l.lunak@collabora.com>
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>